backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
joblib>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
